+++ examples/resources/zabbix_action/import.sh
@@ -1,2 +1,2 @@
-# Import using the Zabbix action ID.
-terraform import zabbix_action.notify_slack 7
+# Import using the Zabbix action ID (look it up in the UI or with the API).
+terraform import zabbix_action.notify_slack "$ZABBIX_ACTION_ID"