+++ examples/resources/zabbix_host_group/import.sh
@@ -1,2 +1,2 @@
-# Import using the Zabbix host group ID.
-terraform import zabbix_host_group.servers 22
+# Import using the Zabbix host group ID (look it up in the UI or with the API).
+terraform import zabbix_host_group.servers "$ZABBIX_HOST_GROUP_ID"