+++ examples/resources/zabbix_media_type/import.sh
@@ -1,2 +1,2 @@
-# Import using the Zabbix media type ID.
-terraform import zabbix_media_type.slack 42
+# Import using the Zabbix media type ID (look it up in the UI or with the API).
+terraform import zabbix_media_type.slack "$ZABBIX_MEDIA_TYPE_ID"